Move `ThemeProvider` to `:core:ui:theme:api` module
