+++ core/ui/theme/api/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id(ThunderbirdPlugins.Library.android)
+}
+
+android {
+    namespace = "app.k9mail.core.ui.theme.api"
+}
