+++ core/ui/theme/api/src/main/kotlin/app/k9mail/core/ui/theme/api/ThemeProvider.kt
@@ -1,4 +1,4 @@
-package app.k9mail.legacy.ui.theme
+package app.k9mail.core.ui.theme.api
 
 import androidx.annotation.StyleRes
