+++ legacy/ui/theme/src/main/java/app/k9mail/legacy/ui/theme/ThemeManager.kt
@@ -5,6 +5,7 @@ import android.content.res.Configuration
 import android.os.Build
 import androidx.annotation.StyleRes
 import androidx.appcompat.app.AppCompatDelegate
+import app.k9mail.core.ui.theme.api.ThemeProvider
 import app.k9mail.legacy.preferences.AppTheme
 import app.k9mail.legacy.preferences.GeneralSettings
 import app.k9mail.legacy.preferences.GeneralSettingsManager
